# api/openapi.yaml
openapi: 3.0.0
info:
  license:
    name: MIT
  title: OurGroceries REST API
  version: v1
servers:
  - url: 'http://grocerer/v1'
paths:
  /lists:
    get:
      operationId: getLists
      summary: Get the shopping lists
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/List'
        '401':
          description: Unauthorized
  /lists/{listID}:
    get:
      operationId: getList
      summary: Get a shopping list
      parameters:
        - name: listID
          in: path
          description: ID of list to get
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '401':
          description: Unauthorized
  /lists/{listID}/items:
    post:
      operationId: addItem
      summary: Add an item to a list
      parameters:
        - name: listID
          in: path
          description: ID of list to add to
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: OurGroceries email
          required: false
          schema:
            type: string
        - name: password
          in: query
          description: OurGroceries password
          required: false
          schema:
            type: string
      requestBody:
        description: Item to add to the grocery list
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        '201':
          description: Item created
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
  /version:
    get:
      operationId: getVersion
      summary: Get the version of the service
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
components:
  schemas:
    Item:
      properties:
        name:
          type: string
          example: "sardines"
    List:
      properties:
        name:
          type: string
          example: "n8w1aAoqqURBrM2nUns6nQ"
        id:
          type: string
          example: "The Shopping List"
    Version:
      properties:
        api:
          type: integer
          example: 1
        commit:
          type: string
          example: '423322c'
        date:
          type: string
          example: '2019-06-17T21:05Z'